% Generated by roxygen2 (4.0.1): do not edit by hand
\name{aggregateObservationFlag}
\alias{aggregateObservationFlag}
\title{This function performs aggregation on the flag}
\usage{
aggregateObservationFlag(..., flagTable = faoswsFlagTable)
}
\arguments{
\item{...}{Vectors of Flags, see detail}

\item{flagTable}{The table which maps the relationship between the
flag and its information weight.}
}
\description{
The aggregation is based on the weight of the flag as designated
by the table. Official sources has high weight while data from
unreliable sources has low weight.
}
\details{
The arguement for flag is the same to the function
sum. Multiple instances of the flag can be pass in to the function
to be aggregated.
}
\seealso{
\code{\link{sum}}
}

