% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkFlagValidity.R
\name{checkFlagValidity}
\alias{checkFlagValidity}
\title{This function checks whether the flags in the data are valid according to
the flagValidTable.}
\usage{
checkFlagValidity(data, flagObservationVar = "flagObservationStatus",
  flagMethodVar = "flagMethod", returnData = TRUE, normalised = TRUE,
  denormalisedKey = "measuredElement", flagTable = flagValidTable)
}
\arguments{
\item{data}{The data to be checked}

\item{flagObservationVar}{The column name corresponding to the observation
status flag.}

\item{flagMethodVar}{The column name corresponding to the method flag.}

\item{returnData}{logical, whether the data should be returned}

\item{normalised}{logical, whether the data is normalised}

\item{denormalisedKey}{optional, only required if the input data is not
normalised.It is the name of the key that denormalises the data.}

\item{flagTable}{The table containing valid/invalid flag combination. See
\data{flagValidTable} in \pkg{faoswsFlag}}
}
\value{
The original data is returned if all flags are valid, otherwise an
    error.
}
\description{
This function checks whether the flags in the data are valid according to
the flagValidTable.
}

