% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combineFlag.R
\name{combineFlag}
\alias{combineFlag}
\title{Function to create a combined flag}
\usage{
combineFlag(data, flagObservationStatusVar, flagMethodVar)
}
\arguments{
\item{data}{The dataset}

\item{flagObservationStatusVar}{The column corresponds to the observation
status flag.}

\item{flagMethodVar}{The column corresponds to the method flag.}
}
\value{
a vector of string with the flags concatenated.
}
\description{
Function to create a combined flag
}

