% Generated by roxygen2 (4.0.1): do not edit by hand
\name{computeFlagWeight}
\alias{computeFlagWeight}
\title{Function to compute the weights of each flag based on the history}
\usage{
computeFlagWeight(history, officialFlag = "", method = c("entropy",
  "similarity"))
}
\arguments{
\item{history}{The history of data queried from the api}

\item{officialFlag}{The character string representing official
figure.}

\item{method}{The method to be used to calculate the weights.}
}
\description{
Function to compute the weights of each flag based on the history
}

