% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{package}
\name{faoswsFlag-package}
\alias{faoswsFlag-package}
\title{Package to manipulate and aggregate observational status
flags.}
\description{
This package provides functions to manipulate and aggregate
observational status flags.
}
\author{
Michael. C. J. Kao \email{michael.kao@fao.org}
}
\keyword{package}

