% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{data}
\name{faoswsFlagTable}
\alias{faoswsFlagTable}
\title{Flag table for the new FAO statistical working system}
\description{
The table maps the relationship between the observational status
flag and it's corresponding information weight
}
\keyword{datasets}

