% Generated by roxygen2 (4.0.1): do not edit by hand
\name{flag2weight}
\alias{flag2weight}
\title{Function to convert flag to information weight.}
\usage{
flag2weight(flagObservationStatus, flagTable = faoswsFlagTable)
}
\arguments{
\item{flagObservationStatus}{The flag}

\item{flagTable}{The table mapping the flag and weights}
}
\description{
The flag is converted to a numeric information weight for
quantification and further manipulation.
}

