% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flagValidTable.R
\docType{data}
\name{flagValidTable}
\alias{flagValidTable}
\title{Flag  validitiy table for the new FAO statistical working system}
\description{
The table determines whether a combination of flags are valid, and at the
same time provide information on whether it is protected against statistical
algorithms.
}
\keyword{datasets}

