% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getProtectedFlag.R
\name{getProtectedFlag}
\alias{getProtectedFlag}
\title{The function obtains the flag combinations which are considered protected.}
\usage{
getProtectedFlag()
}
\description{
The function obtains the flag combinations which are considered protected.
}

