% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getValidFlag.R
\name{getValidFlag}
\alias{getValidFlag}
\title{The function returns the flag combination which are valid}
\usage{
getValidFlag()
}
\description{
The function returns the flag combination which are valid
}

