% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sumObservationFlag.R
\name{sumObservationFlag}
\alias{sumObservationFlag}
\title{This function performs sum over flags}
\usage{
sumObservationFlag(flags, flagTable = flagWeightTable)
}
\arguments{
\item{flagTable}{The table which maps the relationship between the
flag and its information weight. see data{flagWeightTable}}

\item{Vector}{of flags}
}
\value{
A character vector containing the aggregated flags.
}
\description{
The sum is based on the weight of the flag as designated
by the table. Official sources have high weight while data from
unreliable sources have low weight.
}
\examples{
sumObservationFlag(c("T", "E"))
# Passing multiple flags aggregates them all together:
sumObservationFlag(c("", "T", "E", "I", "M"))

}
\seealso{
Similar function \code{\link{aggregateObservationFlag}}
}

