% Generated by roxygen2 (4.0.1): do not edit by hand
\name{weight2flag}
\alias{weight2flag}
\title{Function to convert information weight to flag.}
\usage{
weight2flag(flagObservationWeight, flagTable = faoswsFlagTable)
}
\arguments{
\item{flagObservationWeight}{The flag information weight}

\item{flagTable}{The table mapping the flag and weights}
}
\description{
The function converts a numeric information weight back to flags.
}

