% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkFlagValidity.R
\name{checkFlagValidity}
\alias{checkFlagValidity}
\title{This function checks whether the flags in the data are valid according to
the flagValidTable.}
\usage{
checkFlagValidity(data, flagObservationVar = "flagObservationStatus",
  flagMethodVar = "flagMethod", flagTable = flagValidTable)
}
\arguments{
\item{data}{The data to be checked}

\item{flagObservationVar}{The column name corresponding to the observation
status flag.}

\item{flagMethodVar}{The column name corresponding to the method flag.}

\item{flagTable}{The table containing valid/invalid flag combination.
See \data{flagValidTable} in \pkg{faoswsFlag}}
}
\value{
The original data is returned if all flags are valid, otherwise an
    error.
}
\description{
This function checks whether the flags in the data are valid according to
the flagValidTable.
}

